feat: adjust timezone property for needed classes
