+++ src/Date.ts
@@ -4,19 +4,8 @@ import Field from "./Field";
  * Date input
  */
 class Date extends Field {
-  _timezone: string | undefined = undefined;
-
-  get timezone(): string | undefined {
-    return this._timezone;
-  }
-
-  set timezone(value: string | undefined) {
-    this._timezone = value;
-  }
-
-  constructor(props?: { timezone?: string } & any) {
+  constructor(props?: any) {
     super(props);
-    this._timezone = props?.timezone || undefined;
   }
 }
